+++ src/main/java/com/ecomerceApi/Priscila/config/JwtAuthenticationFilter.java
@@ -1,5 +1,6 @@
 package com.ecomerceApi.Priscila.config;
 
+import com.ecomerceApi.Priscila.service.JwtService;
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
