+++ src/main/java/com/ecomerceApi/Priscila/requestModels/RegisterRequest.java
@@ -1,4 +1,4 @@
-package com.ecomerceApi.Priscila.auth;
+package com.ecomerceApi.Priscila.requestModels;
 
 import com.ecomerceApi.Priscila.model.Role;
 import lombok.AllArgsConstructor;
